TKSS-954: Split kona_crypto.c
