+++ kona-crypto/src/main/jni/kona_sm2.c
@@ -25,564 +25,8 @@
 #include <openssl/core_names.h>
 #include <openssl/ec.h>
 #include <openssl/evp.h>
-#include <openssl/err.h>
 
-#define SM2_PRI_KEY_LEN      32
-#define SM2_PUB_KEY_LEN      65
-#define SM2_COMP_PUB_KEY_LEN 33
-#define SM3_DIGEST_LEN       32
-#define SM3_MAC_LEN          32
-#define SM4_KEY_LEN          16
-#define SM4_GCM_TAG_LEN      16
-
-#define OPENSSL_SUCCESS       1
-#define OPENSSL_FAILURE       0
-#define KONA_GOOD             0
-#define KONA_BAD             -1
-
-#define KONA_print(...) fprintf(stdout, __VA_ARGS__), fprintf(stdout, "\n")
-#define KONA_print_err(...) fprintf(stderr, __VA_ARGS__), fprintf(stderr, "\n")
-#define OPENSSL_print_err() ERR_print_errors_fp(stderr)
-
-const char* hex_digits = "0123456789abcdef";
-void bytes_to_hex(const unsigned char* bytes, size_t offset, size_t len, unsigned char* hex) {
-    for (size_t i = 0; i < len; i++) {
-        hex[i * 2] = hex_digits[bytes[i + offset] / 16];
-        hex[i * 2 + 1] = hex_digits[bytes[i + offset] % 16];
-    }
-
-    hex[len * 2] = '\0';
-}
-
-void print_hex(const unsigned char* byte_array, size_t offset, size_t len) {
-    unsigned char* hex = malloc(len* 2 + 1);
-    bytes_to_hex(byte_array, offset, len, hex);
-    KONA_print("%s", hex);
-    free(hex);
-}
-
-/* ***** SM3 start ***** */
-int sm3_reset(EVP_MD_CTX* ctx) {
-    if (ctx == NULL) {
-        return OPENSSL_FAILURE;
-    }
-
-    if (!EVP_MD_CTX_reset(ctx)) {
-        OPENSSL_print_err();
-        return OPENSSL_FAILURE;
-    }
-
-    const EVP_MD* md = EVP_sm3();
-    if (md == NULL) {
-        OPENSSL_print_err();
-        return OPENSSL_FAILURE;
-    }
-
-    if (!EVP_DigestInit_ex(ctx, md, NULL)) {
-        OPENSSL_print_err();
-        return OPENSSL_FAILURE;
-    }
-
-    return OPENSSL_SUCCESS;
-}
-
-JNIEXPORT jlong JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3CreateCtx
-  (JNIEnv* env, jobject thisObj) {
-    EVP_MD_CTX* ctx = EVP_MD_CTX_new();
-    if (ctx == NULL) {
-        OPENSSL_print_err();
-        return KONA_BAD;
-    }
-
-    const EVP_MD* md = EVP_sm3();
-    if (md == NULL) {
-        OPENSSL_print_err();
-        return KONA_BAD;
-    }
-
-    if (!EVP_DigestInit_ex(ctx, md, NULL)) {
-        OPENSSL_print_err();
-        return KONA_BAD;
-    }
-
-    return (jlong)ctx;
-}
-
-JNIEXPORT void JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3FreeCtx
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_MD_CTX* ctx = (EVP_MD_CTX*)pointer;
-    if (ctx != NULL) {
-        EVP_MD_CTX_free(ctx);
-    }
-}
-
-JNIEXPORT jint JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3Update
-  (JNIEnv* env, jobject thisObj, jlong pointer, jbyteArray data) {
-    EVP_MD_CTX* ctx = (EVP_MD_CTX*)pointer;
-    if (ctx == NULL) {
-      return KONA_BAD;
-    }
-
-    if (data == NULL) {
-        return KONA_BAD;
-    }
-
-    int data_len = (*env)->GetArrayLength(env, data);
-    if (data_len < 0) {
-        return KONA_BAD;
-    }
-
-    jbyte* data_bytes = (*env)->GetByteArrayElements(env, data, NULL);
-    if (data_bytes == NULL) {
-        return KONA_BAD;
-    }
-
-    int result;
-    if (EVP_DigestUpdate(ctx, data_bytes, data_len)) {
-        result = KONA_GOOD;
-    } else {
-        OPENSSL_print_err();
-        result = KONA_BAD;
-    }
-
-    (*env)->ReleaseByteArrayElements(env, data, data_bytes, JNI_ABORT);
-
-    return result;
-}
-
-JNIEXPORT jbyteArray JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3Final
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_MD_CTX* ctx = (EVP_MD_CTX*)pointer;
-    if (ctx == NULL) {
-        return NULL;
-    }
-
-    unsigned char digest[SM3_DIGEST_LEN];
-    unsigned int digest_len = 0;
-
-    if (!EVP_DigestFinal_ex(ctx, digest, &digest_len)) {
-        OPENSSL_print_err();
-        return NULL;
-    }
-
-    if (!sm3_reset(ctx)) {
-        return NULL;
-    }
-
-    if (digest_len <= 0) {
-        return NULL;
-    }
-
-    jbyteArray result = (*env)->NewByteArray(env, digest_len);
-    if (result == NULL) {
-        return NULL;
-    }
-
-    (*env)->SetByteArrayRegion(env, result, 0, digest_len, (jbyte*)digest);
-
-    return result;
-}
-
-JNIEXPORT jint JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3Reset
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_MD_CTX* ctx = (EVP_MD_CTX*)pointer;
-    if (ctx == NULL) {
-        return KONA_BAD;
-    }
-
-    return sm3_reset(ctx) ? KONA_GOOD : KONA_BAD;
-}
-
-JNIEXPORT jlong JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3Clone
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_MD_CTX* orig_ctx = (EVP_MD_CTX*)pointer;
-    if (orig_ctx == NULL) {
-        return KONA_BAD;
-    }
-
-    EVP_MD_CTX* new_ctx = EVP_MD_CTX_new();
-    if (new_ctx == NULL) {
-        return KONA_BAD;
-    }
-
-    if (!EVP_MD_CTX_copy_ex(new_ctx, orig_ctx)) {
-        OPENSSL_print_err();
-        EVP_MD_CTX_free(new_ctx);
-        return KONA_BAD;
-    }
-
-    return (jlong)new_ctx;
-}
-/* ***** SM3 end ***** */
-
-/* ***** SM3HMAC start ***** */
-int sm3hmac_reset(EVP_MAC_CTX* ctx) {
-    if (!EVP_MAC_init(ctx, NULL, 0, NULL)) {
-        OPENSSL_print_err();
-        return OPENSSL_FAILURE;
-    } else {
-        return OPENSSL_SUCCESS;
-    }
-}
-
-JNIEXPORT jlong JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3hmacCreateMac
-  (JNIEnv* env, jobject thisObj) {
-    EVP_MAC* mac = EVP_MAC_fetch(NULL, "HMAC", NULL);
-    if (mac == NULL) {
-        OPENSSL_print_err();
-        return KONA_BAD;
-    }
-
-    return (jlong)mac;
-}
-
-JNIEXPORT void JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3hmacFreeMac
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_MAC* mac = (EVP_MAC*)pointer;
-    if (mac != NULL) {
-        EVP_MAC_free(mac);
-    }
-}
-
-JNIEXPORT jlong JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3hmacCreateCtx
-  (JNIEnv* env, jobject thisObj, jlong macPointer, jbyteArray key) {
-    EVP_MAC* mac = (EVP_MAC*)macPointer;
-    if (mac == NULL) {
-        return KONA_BAD;
-    }
-
-    if (key == NULL) {
-        return KONA_BAD;
-    }
-
-    EVP_MAC_CTX* ctx = EVP_MAC_CTX_new(mac);
-    if (ctx == NULL) {
-        OPENSSL_print_err();
-        return KONA_BAD;
-    }
-
-    OSSL_PARAM params[] = {
-        OSSL_PARAM_construct_utf8_string("digest", "SM3", 0),
-        OSSL_PARAM_construct_end()
-    };
-
-    const int key_len = (*env)->GetArrayLength(env, key);
-    jbyte* key_bytes = (*env)->GetByteArrayElements(env, key, NULL);
-    long result;
-    if (EVP_MAC_init(ctx, (unsigned char*)key_bytes, key_len, params)) {
-        result = (long)ctx;
-    } else {
-        OPENSSL_print_err();
-        result = KONA_BAD;
-    }
-
-    (*env)->ReleaseByteArrayElements(env, key, key_bytes, JNI_ABORT);
-
-    return (jlong)result;
-}
-
-JNIEXPORT void JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3hmacFreeCtx
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_MAC_CTX* ctx = (EVP_MAC_CTX*)pointer;
-    if (ctx != NULL) {
-        EVP_MAC_CTX_free(ctx);
-    }
-}
-
-JNIEXPORT jint JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3hmacUpdate
-  (JNIEnv* env, jobject thisObj, jlong pointer, jbyteArray data) {
-    if (data == NULL) {
-        return KONA_BAD;
-    }
-
-    EVP_MAC_CTX* ctx = (EVP_MAC_CTX*)pointer;
-    if (ctx == NULL) {
-        return KONA_BAD;
-    }
-
-    const int data_len = (*env)->GetArrayLength(env, data);
-    jbyte* data_bytes = (*env)->GetByteArrayElements(env, data, NULL);
-    if (data_bytes == NULL) {
-        return KONA_BAD;
-    }
-
-    int result = KONA_GOOD;
-    if (!EVP_MAC_update(ctx, (unsigned char*)data_bytes, data_len)) {
-        OPENSSL_print_err();
-        result = KONA_BAD;
-    }
-
-    (*env)->ReleaseByteArrayElements(env, data, data_bytes, JNI_ABORT);
-
-    return (jint)result;
-}
-
-JNIEXPORT jbyteArray JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3hmacFinal
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_MAC_CTX* ctx = (EVP_MAC_CTX*)pointer;
-    if (ctx == NULL) {
-        return NULL;
-    }
-
-    unsigned char mac[SM3_MAC_LEN];
-    size_t mac_len = 0;
-
-    if (!EVP_MAC_final(ctx, mac, &mac_len, sizeof(mac))) {
-        OPENSSL_print_err();
-        return NULL;
-    }
-
-    // Reset the context after generating the mac
-    if (!sm3hmac_reset(ctx)) {
-        return NULL;
-    }
-
-    if (mac_len <= 0) {
-        return NULL;
-    }
-
-    jbyteArray result = (*env)->NewByteArray(env, mac_len);
-    if (result == NULL) {
-        return NULL;
-    }
-
-    (*env)->SetByteArrayRegion(env, result, 0, mac_len, (jbyte*)mac);
-
-    return result;
-}
-
-JNIEXPORT jint JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3hmacReset
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_MAC_CTX* ctx = (EVP_MAC_CTX*)pointer;
-    if (ctx == NULL) {
-        return KONA_BAD;
-    }
-
-    return sm3hmac_reset(ctx) ? KONA_GOOD : KONA_BAD;
-}
-
-JNIEXPORT jlong JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm3hmacClone
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_MAC_CTX* orig_ctx = (EVP_MAC_CTX*)pointer;
-    if (orig_ctx == NULL) {
-        return KONA_BAD;
-    }
-
-    EVP_MAC_CTX* new_ctx = EVP_MAC_CTX_dup(orig_ctx);
-    if (new_ctx == NULL) {
-        OPENSSL_print_err();
-        return KONA_BAD;
-    }
-
-    return (jlong)new_ctx;
-}
-/* ***** SM3HMAC end ***** */
-
-/* ***** SM4 start ***** */
-JNIEXPORT jlong JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm4CreateCtx
-  (JNIEnv* env, jobject thisObj, jboolean encrypt, jstring mode, jboolean padding, jbyteArray key, jbyteArray iv) {
-    if (key == NULL) {
-        return KONA_BAD;
-    }
-
-    const char* mode_str = (*env)->GetStringUTFChars(env, mode, 0);
-    const char* sm4_mode = NULL;
-
-    if (strcmp(mode_str, "ECB") == 0) {
-        sm4_mode = "SM4-ECB";
-    } else if (strcmp(mode_str, "CBC") == 0) {
-        sm4_mode = "SM4-CBC";
-    } else if (strcmp(mode_str, "CTR") == 0) {
-        sm4_mode = "SM4-CTR";
-    } else if (strcmp(mode_str, "GCM") == 0) {
-        sm4_mode = "SM4-GCM";
-    } else {
-        (*env)->ReleaseStringUTFChars(env, mode, mode_str);
-        return KONA_BAD;
-    }
-
-    const EVP_CIPHER* cipher = EVP_CIPHER_fetch(NULL, sm4_mode, NULL);
-    if (cipher == NULL) {
-        OPENSSL_print_err();
-        (*env)->ReleaseStringUTFChars(env, mode, mode_str);
-        return KONA_BAD;
-    }
-
-    EVP_CIPHER_CTX* ctx = EVP_CIPHER_CTX_new();
-    if (ctx == NULL) {
-        OPENSSL_print_err();
-        (*env)->ReleaseStringUTFChars(env, mode, mode_str);
-        return KONA_BAD;
-    }
-
-    jbyte* key_bytes = (*env)->GetByteArrayElements(env, key, NULL);
-    jbyte* iv_bytes = iv ? (*env)->GetByteArrayElements(env, iv, NULL) : NULL;
-
-    jlong result = KONA_BAD;
-    if (EVP_CipherInit_ex(ctx, cipher, NULL, (unsigned char*)key_bytes, (unsigned char*)iv_bytes, encrypt)) {
-        if (!padding && !EVP_CIPHER_CTX_set_padding(ctx, 0)) {
-            OPENSSL_print_err();
-        } else {
-            result = (jlong)ctx;
-        }
-    } else {
-        OPENSSL_print_err();
-    }
-
-    (*env)->ReleaseStringUTFChars(env, mode, mode_str);
-    (*env)->ReleaseByteArrayElements(env, key, key_bytes, JNI_ABORT);
-    if (iv_bytes) {
-        (*env)->ReleaseByteArrayElements(env, iv, iv_bytes, JNI_ABORT);
-    }
-
-    if (result == KONA_BAD && ctx != NULL) {
-        EVP_CIPHER_CTX_free(ctx);
-    }
-
-    return result;
-}
-
-JNIEXPORT void JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm4FreeCtx
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_CIPHER_CTX* ctx = (EVP_CIPHER_CTX*)pointer;
-    if (ctx != NULL) {
-        EVP_CIPHER_CTX_free(ctx);
-    }
-}
-
-JNIEXPORT jbyteArray JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm4Update
-  (JNIEnv* env, jobject thisObj, jlong pointer, jbyteArray in) {
-    EVP_CIPHER_CTX* ctx = (EVP_CIPHER_CTX*)pointer;
-    if (ctx == NULL) {
-        return NULL;
-    }
-
-    jbyte* in_bytes = (*env)->GetByteArrayElements(env, in, NULL);
-    if (in_bytes == NULL) {
-        return NULL;
-    }
-    jsize in_len = (*env)->GetArrayLength(env, in);
-
-    int out_len = in_len + EVP_CIPHER_CTX_block_size(ctx);
-    unsigned char* out_buf = (unsigned char*)malloc(out_len);
-    if (out_buf == NULL) {
-        (*env)->ReleaseByteArrayElements(env, in, in_bytes, JNI_ABORT);
-        return NULL;
-    }
-
-    jbyteArray out_bytes = NULL;
-    int len = 0;
-    if (EVP_CipherUpdate(ctx, out_buf, &len, (unsigned char*)in_bytes, in_len)) {
-        out_bytes = (*env)->NewByteArray(env, len);
-        if (out_bytes != NULL && len > 0) {
-            (*env)->SetByteArrayRegion(env, out_bytes, 0, len, (jbyte*)out_buf);
-        }
-    } else {
-        OPENSSL_print_err();
-    }
-
-    (*env)->ReleaseByteArrayElements(env, in, in_bytes, JNI_ABORT);
-    free(out_buf);
-
-    return out_bytes;
-}
-
-JNIEXPORT jbyteArray JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm4Final
-  (JNIEnv* env, jobject thisObj, jlong pointer) {
-    EVP_CIPHER_CTX* ctx = (EVP_CIPHER_CTX*)pointer;
-    if (ctx == NULL) {
-        return NULL;
-    }
-
-    int block_size = EVP_CIPHER_CTX_block_size(ctx);
-    unsigned char* out_buf = (unsigned char*)malloc(block_size);
-    if (out_buf == NULL) {
-        return NULL;
-    }
-
-    jbyteArray out_bytes = NULL;
-    int len = 0;
-    if (EVP_CipherFinal_ex(ctx, out_buf, &len)) {
-        out_bytes = (*env)->NewByteArray(env, len);
-        if (out_bytes != NULL && len > 0) {
-            (*env)->SetByteArrayRegion(env, out_bytes, 0, len, (jbyte*)out_buf);
-        }
-    } else {
-        OPENSSL_print_err();
-    }
-
-    free(out_buf);
-
-    return out_bytes;
-}
-
-JNIEXPORT jint JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm4GCMUpdateAAD
-  (JNIEnv* env, jobject thisObj, jlong pointer, jbyteArray aad) {
-    EVP_CIPHER_CTX* ctx = (EVP_CIPHER_CTX*)pointer;
-    if (ctx == NULL || aad == NULL) {
-        return KONA_BAD;
-    }
-
-    int aad_len = (*env)->GetArrayLength(env, aad);
-    if (aad_len <= 0) {
-        return KONA_BAD;
-    }
-
-    jbyte* aad_bytes = (*env)->GetByteArrayElements(env, aad, NULL);
-    if (aad_bytes == NULL) {
-        return KONA_BAD;
-    }
-
-    int out_len = 0;
-    int result = KONA_BAD;
-    if (EVP_CipherUpdate(ctx, NULL, &out_len, (unsigned char*)aad_bytes, aad_len)) {
-        result = KONA_GOOD;
-    } else {
-        OPENSSL_print_err();
-    }
-
-    (*env)->ReleaseByteArrayElements(env, aad, aad_bytes, JNI_ABORT);
-
-    return result;
-}
-
-JNIEXPORT jint JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm4GCMProcTag
-  (JNIEnv* env, jobject thisObj, jlong pointer, jbyteArray tag) {
-    EVP_CIPHER_CTX* ctx = (EVP_CIPHER_CTX*)pointer;
-    if (ctx == NULL || tag == NULL) {
-        return KONA_BAD;
-    }
-
-    int tag_len = (*env)->GetArrayLength(env, tag);
-    if (tag_len != SM4_GCM_TAG_LEN) {
-        return KONA_BAD;
-    }
-
-    int result = KONA_BAD;
-    if (EVP_CIPHER_CTX_encrypting(ctx)) {
-        unsigned char tag_buf[SM4_GCM_TAG_LEN];
-        if (EVP_CIPHER_CTX_ctrl(ctx, EVP_CTRL_GCM_GET_TAG, SM4_GCM_TAG_LEN, tag_buf)) {
-            (*env)->SetByteArrayRegion(env, tag, 0, SM4_GCM_TAG_LEN, (jbyte*)tag_buf);
-            result = KONA_GOOD;
-        } else {
-            OPENSSL_print_err();
-        }
-    } else {
-        jbyte* tag_bytes = (*env)->GetByteArrayElements(env, tag, NULL);
-        if (tag_bytes != NULL) {
-            if (EVP_CIPHER_CTX_ctrl(ctx, EVP_CTRL_GCM_SET_TAG, SM4_GCM_TAG_LEN, tag_bytes)) {
-                result = KONA_GOOD;
-            } else {
-                OPENSSL_print_err();
-            }
-
-            (*env)->ReleaseByteArrayElements(env, tag, tag_bytes, JNI_ABORT);
-        }
-    }
-
-    return result;
-}
-/* ***** SM4 end ***** */
+#include "kona_common.h"
 
 /* ***** SM2 MISC start ***** */
 JNIEXPORT jbyteArray JNICALL Java_com_tencent_kona_crypto_provider_nativeImpl_NativeCrypto_sm2ToUncompPubKey
